refactor(file structure): rename folder
