--- src/features/categories-list/index.ts
@@ -1,1 +1,0 @@
-export { PaymentsFlatList } from './categories-flat-list'
