+++ src/features/payments-list/index.ts
@@ -1,0 +1,1 @@
+export { PaymentsFlatList } from './payments-flat-list'
