+++ src/pages/payments/servieces-list/ui.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable no-unused-vars */
 import React from 'react'
 import { styled } from '@ui/theme'
-import { PaymentsFlatList } from '@features/categories-list'
+import { PaymentsFlatList } from '@features/payments-list'
 import { PaymentsStackParamList } from '@processes/routing/types'
 import { SearchBar } from '@shared/ui/molecules'
 import { PaymentServiceUI } from '@shared/api/payment-categories'
